♻️ Extract CachePolicy duplication into a __using__ block
